refactored plugin
